:bento: Update PCB to set correct resistor values
In #20 the resistor values have been changed in the schematic.

As this was no PCB change, I did not mind to update the PCB file. However,
the old resistor values are now shown there and also in the BOM.

This is only an update to the resistor values stored in the PCB file, no changes
to the actual PCB layout have been made.