:wrench: Revert input resistors to 10K
Test built with SMD LEDs shows that the 100K base resistor does not
open the transistors sufficiently. Change back to 10K for robust
operation.

The problem that occurred: With lower LED series resistors (220R) the
transistor would be part of a voltage divider leading to 6V output,
which is not enough to trigger the 4024 clock input.